drafts

+++ backend/msgs/utils.py
@@ -5,12 +5,14 @@ def new_event_msgs(reference):
     '''
         Create new msgs row in the database, related to a specific event form
     '''
-    event_form_msgs = Msgs(reference=reference)
-    event_form_msgs.save()
+    if not Msgs.objects.filter(reference=reference).exists():
+        event_form_msgs = Msgs(reference=reference)
+        event_form_msgs.save()
 
 def delete_event_messages(reference):
     '''
         Delete specific event form msgs row in the database 
     '''
-    event_form_msgs = Msgs.objects.get(reference=reference)
-    event_form_msgs.delete()
+    if Msgs.objects.filter(reference=reference).exists():
+        event_form_msgs = Msgs.objects.get(reference=reference)
+        event_form_msgs.delete()
